moved StlPrettyPrint to Utilities
